+++ src/app/blog/[slug]/page.tsx
@@ -4,7 +4,7 @@ import remarkGfm from "remark-gfm";
 import rehypeSlug from "rehype-slug";
 import rehypePrettyCode from "rehype-pretty-code";
 import type { Options as PrettyCodeOptions } from "rehype-pretty-code";
-import { addLazyLoading } from "@/lib/rehype-lazy-loading";
+import { addLazyLoading } from "@/lib/mdx-lazy-loading";
 import { postService } from "@/lib/container";
 import { extractHeadings } from "@/lib/toc";
 import { DATE_FORMAT } from "@/lib/constants";
